This commit has been tested and seems to work with the summer of 2011 database.
I added some dependencies in egg info. If it doesent work, the extra packages needed are

* MySQL-python
* tw.dynforms
* tw.tinymce

Also, *this is uggly* we use european style dates throughout so I had to manual hack the
lib/python-2.5/site-packages/....FormEncode...validators.py and that change hasnt been commited yet (or submitted to the authors yet).

Finally, the dojo lib has to be included in the public folder, and dojo 1.5 is naturally not tracked by this repo.

So there are some rough edges.